requirements.txt
Flask~=2.3.2
keras~=2.12.0
tensorflow~=2.12.0
numpy~=1.23.5
Pillow~=9.5.0
gevent~=22.10.2
